api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
passlib[bcrypt]==1.7.4
pydantic==2.10.6
python-dotenv==1.0.1
